requirements.txt

# Basic Python dependencies
python-jose[cryptography]==3.3.0
bcrypt==3.2.0
requests==2.25.1
six==1.16.0
jsonschema==3.2.0
Jinja2

# ai/sci tools
numpy==1.22.3
pytesseract==0.3.9
opencv-python==4.5.5.64
decord==0.6.0
scikit-image==0.19.2
mtcnn-opencv==1.0.2
nltk==3.7
prefixspan==0.5.2

# Flask + Connexion + OpenAPI dependencies
Flask==2.0.2
Werkzeug==2.0.1
connexion[swagger-ui]==2.7.0
prance==0.21.8.0

# Database plugins for Flask
flask_sqlalchemy==2.5.1
marshmallow-sqlalchemy==0.27.0
marshmallow==3.14.1
flask-marshmallow==0.14.0
SQLAlchemy==1.4.31
psycopg2-binary==2.9.3

# RabbitMQ
pika==1.2.0
pika-stubs==0.1.3
rabbitpy==2.0.1

# ClassTranscribe python deps
grpcio-tools
jupyter
scenedetect[opencv,progress_bar]
KalturaApiClient

pytesseract==0.3.9
